refactor: calendar modal
